Enable credential passing for client
The client works great for me, I love how intuitive it is to use. Since my mqtt server requires authentication and I saw no builtin way to use it I added it this way to the client startup.

In paho one calls _client.set_username_pw after the client is instantiated but with the context manager setting up the context fails. Because one never enterst the context one never gets a chance to pass the credentials to the client.