fix: enable python3.6 support
Client uses contextlib.asynccontextmanager, was added in python3.7 and is
the only thing that prevents python3.6 from working.  Luckily, there is a
backport of that functionality in async_generator.asynccontextmanager.
Conditionally load the async_generator if missing from contextlib and change
setup to require async_generator depending on the python version.

Signed-off-by: Derrick Lyndon Pallas <derrick@meter.com>