client: Fix thread safety in socket close callback
The `_on_socket_close` method can be called from a different thread than the
running event loop if an error occurs while the connection is being
established.  This can happen, for example, if the TCP connection is
established, but the TLS or WebSocket handshake fails.  It is not safe to
cancel the task from another thread.  Instead, the cancellation should be
scheduled to run as soon as the event loop is idle.