fix: consistently use a user-provided logger
Client initializer currently accepts a logger as an optional argument.
That logger is passed to the underlying paho.mqtt library and then used
for logging there. However, there also exists MQTT_LOGGER, which is used
internally by asyncio-mqtt. Without looking at the source code it is
easy to forget that in order to enable logging both in paho.mqtt and in
asyncio-mqtt it is not enough to pass a logger as an argument to Client
initializer.

This commit makes asyncio-mqtt use a user-provided logger instead of
MQTT_LOGGER for internal logging, so that if user decides to pass a
logger to Client constructor, that logger will be used both in paho.mqtt
and asyncio-mqtt.